Feature/ljh/works detail

+++ server/seed/index.ts
@@ -1,22 +1,21 @@
 import fs from 'fs';
 import path from 'path';
-import User from '../models/user';
-import users from './user';
+
+const connect = require('../config/mongo');
 
 require('dotenv').config();
-const connect = require('../config/mongo');
 
 const basename = path.basename(__filename);
 
 connect();
 fs.readdirSync(path.join(__dirname))
   .forEach(async (file) => {
-    if (file !== basename) {
+    if (file !== basename && file !== 'ids.ts') {
       const data = require(path.join(__dirname, file)).default;
       const base = path.basename(file);
       const model = require(path.join(__dirname, '../models', base)).default;
       await model.deleteMany({});
-      model.create(data, (err) => {
+      await model.create(data, (err) => {
         console.error('error: ', err);
         process.exit(0);
       });
